requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
pydantic>=2.5.0
uuid6>=1.0.0
pytest>=7.0.0
